fix: separate Modal logic into its own class and add wrapper View tag to Modal
Separate the logic for the modal and its text in components/PopupModal.js

Wrap the Modal tag with a View tag.

This fixes an issue with the Map freezing after closing a Modal.
Now, the Modal slides from the bottom to the middle of the map, and Map functionality resumes.