Modal has been created
Modal has been created as a separate component.
Efforts have been started to make the borders of the button flash to indicate that this is the button: that button -- the button that needs to be pushed ... pushed to cause the ... cause the transition ... transition to ... transition to the other campus.