Wrapping up the linking of Directions buttons
Refactored GetDirections so that it could be opened from both the pop up and the navbar. I included a boolean value that fromPopup. In case if it was true, it would set the destination to the location of the pressed marker, else the destiantion would stay false.
created a couple configuration files : namely, jest.confi.js and react-native-webview.js, because WebView couldn't.
 be located my expo cli.
I have refactored the testing suits for MapMarkers, GetDireciton, and PopupModal because the source code had been refactored prior.