Linking Multistep navigation to Calendar, fetching and contextualizing current location
merged develop and the calendar branch.

added polygons for points of interests to determine if the user is within a certain building that we know of.

Added the funcitonality to determine, if the user is in a building we have records of: Concordia or Point Of Interst.

Added the funcitonality to fetch the address of the class from the calendar event. It is conditional tho because it depends on the user to enter it.

The functionality isn't fully completed

Added coverage to gitignore:
- After merging develop into this branch, the coverage got erased, so here I am, adding it for the second time